Add game specific sprites
