+++ types.ts
@@ -123,10 +123,14 @@ export class Games {
 		},
 	];
 
-	static findEntry = (game: string): VersionGroup | undefined => {
-		return this.names.find(
+	static findEntry = (game: string): VersionGroup => {
+		const result = this.names.find(
 			a => a.version_group_name === game || a.consistsOf.includes(game)
 		);
+
+		return result
+			? result
+			: { version_group_name: "all", consistsOf: [], generation: "all" };
 	};
 }
 
@@ -203,20 +207,21 @@ export interface Evolution {
 	targetURL: string;
 	targetSprite: string;
 }
-export interface PokemonDetails extends PokemonName {
+export interface PokemonDetails {
+	german: string;
+	english: string;
+	id: number;
+	link: string;
 	types: {
 		name: string;
 		sprite: string;
 	}[];
 	forms: {
 		name: string;
-		sprite: string;
-		spriteShiny: string;
+		sprites: GenericSprites;
 		url: string;
 	}[];
-	backSprite: string;
-	shinySprite: string;
-	shinyBackSprite: string;
+	sprites: GenericSprites;
 	captureRate: number;
 	growthRate: string;
 	evolutions: Evolution[];
@@ -251,15 +256,18 @@ export interface APIResponseForm {
 		};
 		name: string;
 	}[];
-	sprites: {
-		front_default: string;
-		front_shiny: string;
-		front_female: string;
-		front_shiny_female: string;
-	};
+	sprites: SpriteForm;
+}
+
+export interface SpriteForm {
+	front_default: string;
+	front_shiny: string;
+	front_female: string;
+	front_shiny_female: string;
 }
 
 export interface APIResponsePokemon {
+	id: number;
 	weight: number;
 	height: number;
 	abilities: {
@@ -312,14 +320,180 @@ export interface APIResponsePokemon {
 			name: string;
 		};
 	}[];
-	sprites: {
-		back_default: string;
-		back_shiny: string;
-		front_default: string;
-		front_shiny: string;
+	sprites: Sprites;
+}
+
+export interface Sprites {
+	back_default: string;
+	back_shiny: string;
+	front_default: string;
+	front_shiny: string;
+	other: {
+		home: {
+			front_default: string;
+			front_female: string;
+			front_shiny: string;
+			front_shiny_female: string;
+		};
+		"official-artwork": {
+			front_default: string;
+			front_shiny: string;
+		};
+	};
+	versions: {
+		"generation-i": {
+			"red-blue": {
+				back_default: string;
+				back_gray: string;
+				back_transparent: string;
+				front_default: string;
+				front_gray: string;
+				front_transparent: string;
+			};
+			yellow: {
+				back_default: string;
+				back_gray: string;
+				back_transparent: string;
+				front_default: string;
+				front_gray: string;
+				front_transparent: string;
+			};
+		};
+		"generation-ii": {
+			crystal: {
+				back_default: string;
+				back_shiny: string;
+				back_shiny_transparent: string;
+				back_transparent: string;
+				front_default: string;
+				front_shiny: string;
+				front_shiny_transparent: string;
+				front_transparent: string;
+			};
+			gold: {
+				back_default: string;
+				back_shiny: string;
+				front_default: string;
+				front_shiny: string;
+				front_transparent: string;
+			};
+			silver: {
+				back_default: string;
+				back_shiny: string;
+				front_default: string;
+				front_shiny: string;
+				front_transparent: string;
+			};
+		};
+		"generation-iii": {
+			emerald: { front_default: string; front_shiny: string };
+			"firered-leafgreen": {
+				back_default: string;
+				back_shiny: string;
+				front_default: string;
+				front_shiny: string;
+			};
+			"ruby-sapphire": {
+				back_default: string;
+				back_shiny: string;
+				front_default: string;
+				front_shiny: string;
+			};
+		};
+		"generation-iv": {
+			"diamond-pearl": {
+				back_default: string;
+				back_female: string;
+				back_shiny: string;
+				back_shiny_female: string;
+				front_default: string;
+				front_female: string;
+				front_shiny: string;
+				front_shiny_female: string;
+			};
+			"heartgold-soulsilver": {
+				back_default: string;
+				back_female: string;
+				back_shiny: string;
+				back_shiny_female: string;
+				front_default: string;
+				front_female: string;
+				front_shiny: string;
+				front_shiny_female: string;
+			};
+			platinum: {
+				back_default: string;
+				back_female: string;
+				back_shiny: string;
+				back_shiny_female: string;
+				front_default: string;
+				front_female: string;
+				front_shiny: string;
+				front_shiny_female: string;
+			};
+		};
+		"generation-v": {
+			"black-white": {
+				animated: {
+					back_default: string;
+					back_female: string;
+					back_shiny: string;
+					back_shiny_female: string;
+					front_default: string;
+					front_female: string;
+					front_shiny: string;
+					front_shiny_female: string;
+				};
+				back_default: string;
+				back_female: string;
+				back_shiny: string;
+				back_shiny_female: string;
+				front_default: string;
+				front_female: string;
+				front_shiny: string;
+				front_shiny_female: string;
+			};
+		};
+		"generation-vi": {
+			"omegaruby-alphasapphire": {
+				front_default: string;
+				front_female: string;
+				front_shiny: string;
+				front_shiny_female: string;
+			};
+			"x-y": {
+				front_default: string;
+				front_female: string;
+				front_shiny: string;
+				front_shiny_female: string;
+			};
+		};
+		"generation-vii": {
+			icons: {
+				front_default: string;
+				front_female: string;
+			};
+			"ultra-sun-ultra-moon": {
+				front_default: string;
+				front_female: string;
+				front_shiny: string;
+				front_shiny_female: string;
+			};
+		};
+		"generation-viii": {
+			icons: {
+				front_default: string;
+				front_female: string;
+			};
+		};
 	};
 }
 
+export interface GenericSprites {
+	primary: string;
+	secondary: string;
+	shiny: string;
+}
 export interface APIResponseAbility {
 	name: string;
 	names: {
